virtc: Add volume parameters to command line
This patch adds one new -volume parameter to the command line.
It is used to define which mount tags and host paths should be
added as virtual devices to get the agent working properly.
It expects a string list delimited with " ". Each string should
contain ':' delimiter to split mount tag and host path parameters.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>