agent: Check connection status every time start() is called
We want our agent implementation to independently check if the
connection with the agent on the guest is still up. This means
that we want to reconnect in case we lost the connection or to
do nothing in case we can still communicate.

From pod.go perspective, we just want to call start() from
agent interface, without wondering bothering about the agent
connection state.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>