agent: Fix hanging connection issues
Sometimes when trying to stop the pod after it has been previously
started, we were hanging because the agent was trying to connect to
hyperstart but the previous connection was not closed properly.
Because we don't want to open/close the connection every time we
call an agent function interacting with agent running inside the VM,
the best way is to indicate from the API level that we know nothing
else will come after the command endSession, meaning we can properly
close what have been previously opened.
It doesn't break the ability to detect/reconnect during the same
session if we lost the connection for any reason.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>