api: Add proper locking to container API
We don't want two concurrent processes to be able to access the same
pod at the same time.
In case of Create and Delete functions, this means we want to make
sure we are not modifying pod configuration at the same time.
In case of Start, Stop and Enter, we don't want to start several
connections with the agent on the same VM. It would result in non
expected behaviors.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>