pod: Fix defer calls in pod unit testing
This commit fixes an issue related to os.Exit(). Indeed, when we call
this system call at the end of a golang function, there is no way our
deferred calls will be executed. This patch intends to fix that issue.

Also we need to create the testBundle in order to be accessible from
any unit test later.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>