hyperstart: mock: Enhance mock behavior
In case of the command DESTROYPOD, we won't receive anything from
hyperstart running inside the VM, meaning we want the same behavior
with the mock.

Also, we don't want to assert an error when the writeCtl fails because
the connection has been closed from client side, because it is not an
error. We just need to handle this and stop sending anything to the client.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>