network: qemu: Add network endpoints to hypervisor
This patch modifies the hypervisor interface about createPod(), in order
to pass network endpoints as parameters. This will be used by the hypervisor
implementation to add the network pairs to its specific config.

This patch implements qemu to be able to add network devices according to
the network pairs passed as parameters.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>