Introduce pause container
The hyperstart protocol and APIs changed recently and it came
with a change related to pod lifecycle. Now when a pod has not
any workload running, the pod is killed and there is no way to
start a new container unless we recreate this pod.

This patch does several things to make sure we can properly
manage pod and container lifecycle:
- Ability to start a pod with N container
- Mount/Unmount properly containers rootfs and pause binary
- Start a "pause" container on the pod, preventing the pod from terminating

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>