virtc: Modify available parameters related to pause container
This patch removes two parameters for virtc CLI:
bundle: we don't want to specify a rootfs for the pod itself, it will
be specified for the container.
init-cmd: we don't want to specify a command to be executed at the
pod creation, it will be specified for the container.

It also adds a new parameter pause-path, in order to provide
the path containing the pause binary.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>