network: cni: Add unit testing
CNI package handling the usage of the CNI plugin was not covered by
any unit test until now. This patch aims to cover as much as possible
the unit testing of this cni package.

Furthermore, some minor changes have been brought to the cni package
itself in order to be able to test the package more easily.

Notice that we need to modify travis.yml to properly setup the
test environment.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>