agent: hyperstart: Remove DESTROYPOD
In some cases, we noticed some issues with the VM not being terminated
properly from the hypervisor because the VM itself was already down.
This is caused by the DESTROYPOD command which removes a pod and also
kill the init process, resulting in the end of the VM.

Because stopping a pod corresponds to end the VM from the hypervisor,
we don't need to specify DESTROYPOD from hyperstart protocol, we can
only remove all the containers and then let the hypervisor kill the VM.

That's why this commit is removing the DESTROYPOD message from the
stopPod() agent implementation of hyperstart.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>