api_test: Update unit testing because of new Proxy interface
Now that we have the Proxy interface, and now that we don't have a
raw implementation for the hyperstart protocol, we don't need to use
the hyperstart mock anymore.
Because the proxy will change in a near future, we don't create a mock
cc-proxy here, we only rely on the noopProxy implementation.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>