hook: Handle hooks specified by OCI specification
In case of OCI specification we can expect to receive some hooks
configurations that we have to handle. For example, those hooks
can be really important to setup the network between the host
and the network namespace.

Refer to https://github.com/opencontainers/runtime-spec/blob/master/config.md#hooks
for the full OCI hooks specification.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>