cni: Don't switch to netns when creating network
CNI network implementation does not need to switch completely to the
specified netns at the time of the add(). It only needs to execute the
creation and linking of bridge and TAP interfaces inside this netns, but
then it can return to the initial netns.

The join() function will be called from StartPod() to switch permanently
to the specified netns.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>