network: Move network creation to Start and deletion to Stop
Because of the way Docker is working and the OCI specs is defined, we cannot
set the network as is. Indeed, we have to move the network creation to the
StartPod() function and the network deletion to StopPod() function.

The way the CNM hook works, it cannot setup the network if the pod is not
started before. Thus, we cannot add the network during the pod creation.
And in order to be symmetrical and consistent with that, we decide to
perform the removal of the network when the pod is stopped instead of
when it is deleted.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>