hyperstart: Add multicast support to handle PAE messages
Recent hyperstart changes force us to redesign the way our hyperstart
package handle the communication on the CTL channel. Indeed, we now
have to handle synchronized and asynchronized on the same channel.

This commit introduces a multicaster to listen and route the messages
to internal Go channels inside the package.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>