hyperstart: Update package based on multicast support
Now that multicast have been introduced, we make the needed changes
to our existing package APIs.
This commit relies on previous changes adding the multicast support
so that we can receive PROCESSASYNCEVENT in an asynchronous way, and
in the same time being able to handle the synchronous messages we can
expect to receive after we sent a command.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>