hook: Add unit testing for OCI hooks
In order to perform a proper unit testing of the hooks, we cannot rely
on Docker complex setup. Instead, this patch adds a new binary mocking
the behavior of dockerd binary.
It basically checks that we receive expected arguments and the right
numbers of them. Then it reads on stdin to retrieve a json frame which
corresponds to the expected HookState structure.

Unit tests rely on this mock binary instead of the real dockerd.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>