network: Fix the use of network namespaces
We have recently realized it was not correct to switch our StartPod()
function to a specific namespace without getting back to the original
one. Indeed, switching to a namespace without protecting the call with
runtime.LockOSThread() will potentially end with another thread not
being executed in the same namespace.
To ensure we are not mixing network namespaces, we have identified
specific places in the code where we need to execute some actions
in a specific network namespace.

In order to achieve that properly, this patch modifies the join() method
from network interface, so that we can use it to execute anything inside
the given netns.

Once this new method has been implemented, all the places where we needed
to be in the given netns have been encapsulated inside the join() function.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>