api: Remove the network after PostStopHooks run
This patch fixes the sequence for the StopPod() API. We don't
want to remove the network setup and its namespace before to
execute post stop hooks. Those hooks has to be started from
the network namespace.

The patch also moves the part where we fetch pod network data
to an earlier stage.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>