qemu: Fix netns issues due to race conditions
The issue we have encountered was to make sure we didn't have any concurrent calls to the
netNS.Do() function because it allows to execute some functions inside a given network
namespace but it has to run in a single thread.
We were running our qemu VM from a go routine and we were setting the netns inside this
go routine. The goal is to make sure that qemu is started and in case it is not, we can
return an error due to a timeout. But this approach needs to be done in a very specific
way to avoid conflicts with other potential Do() calls. We have to make sure the go routine
is done at the moment it returns a message on our internal channel. That way we can respect
the single thread constraint regarding netNS.Do() calls.

Because all of what was done after the message to notify the VM was started was sent, was not
used, this patch removes everything related to qmpCommand.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>