api: Move VM and network creation/deletion to CreatePod/DeletePod
Because we want to be more efficient and save VM boot performances, we
decided to move the VM creation to the CreatePod() function. This means
the network creation has also been moved to the same function.
In order to keep this symetrical, the network removal and the VM
deletion are now done inside DeletePod().

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>