pod: Add token field to Cmd structure
Runtime implementation relying on virtcontainers has to be able to get the
shim token returned by the proxy to our agent implementation. Because of this
requirement, we consider a token/shim can be started for every process run on
any container, meaning each Cmd structure we fill, will be used to return the
token associated to this process.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>