container: Kill a container before to remove it
Removing a container can fail because the process inside the container is
still up and running. To prevent this case to happen, when doing a stop of
the container, we now first kill it, and then remove it.

Because killing the container also removes it (hyperstart behavior), we need
to be permissive regarding the return of container removal and only log the
error instead of returning it.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>