tests: Move TestMain() to new test file and make testDir a variable.
There can only be a single TestMain() function per package. Since there
are other test files that need a TestMain() in the virtcontainers
package, move the function to its own test file to make it clear that it
is not tied to pod_test.go.

Also, move some consts and crucially the package variables (including
testDir) to this new test file to allow TestMain() to set the values and
ensure all paths are using temporary directories, rather than host-level
directories. This ensures the tests don't impact the host system and
also allow more tests to run as non-root.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>