hyperstart: Move VM registration to startAgent()
Runtime implementations such as OCI can expect to retrieve a PID
corresponding to the running process at the time they call into
create() function.
In our case, this means that we have to provide the PID of the shim
we have just started. But we also expect to get a token from the proxy
before we can start the shim, otherwise it won't be authorized by the
proxy. That is why we move the registration of the VM to the function
startAgent() which is called by the CreatePod() API.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>