agent: Add createContainer to store ProxyInfo
In order to mirror what we have done with the pod, that is having
startAgent() getting information from proxy at the create time, we
want the same for containers. That's why we introduce a new function
createContainer in the agent interface, called at container creation
and responsible for asking the proxy about all information it can
provide.

Another reason for having this patch is to be able to create a
container between pod creation and pod starting. Indeed, we cannot
currently create a container from the API exposed if it occurs before
the pod is started. The reason is that no sequence IDs are issued for
this new container, and hyperstart will send an error because the call
to agent.startPod() won't be able to provide sequence IDs for this new
container.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>