agent: Cleanup the interface using consistent parameters
Functions exposed by the agent interface are a bit confusing regarding
their parameters. Indeed, we didn't really know if we could expect
them to be modified or not. Now we make sure that we pass a pointer
in case we expect some modifications from the agent, otherwise we
only give a copy of the object used.

By removing the Pod pointer from the hyperstart agent instance, we
remove the confusion with which pointer will be used and make sure
it is up to date.

Moreover, we replaced all the ContainerConfig parameters with some
Container ones, because we have more information inside the object
Container, including the information about the configuration. It is
particularly useful to get and set the Token related to a container
process.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>