pod: Retrieve pod URL and convey it through pod structure
Runtime implementation using virtcontainers expects to get a URL
corresponding to the entry point to discuss with the proxy. This
URL is provided by the proxy itself through calls to register()
and attach().
In order to achieve this, we had to change a bit the agent interface
and the API exposed by virtcontainers.

With this change, a runtime can now rely on the Pod handle returned
to retrieve the URL that it can pass to its shim.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>