pod: Enable pod console traces
We had no way to get logs printed inside the VM by our hyperstart
agent because we were missing several parts to be connected properly.

This patch enables the proxy to retrieve hyperstart logs coming from
inside the VM through a new console socket created by the hypervisor.

First it adds a new Console field to the PodConfig so as to be able
to share this socket path with other components such as proxy or
agent implementations.

Then we have to change systemd.unit parameter from "container.target"
to "cc-agent.target" because we need hyperstart output to be
redirected to the right console.

At last, the hypervisor implementation has to create a console socket
from the console path provided by the PodConfig.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>