network: Remove ciao/networking/libsnnet dependency
Because we need to use netlink Go package, and because libsnnet also
relies on this same package, we cannot have both packages at the same
time. It is caused by the vendoring in ciao while we are following the
upstream in virtcontainers.

This patch removes all dependencies related to libsnnet package.
Indeed, the netlink package is really needed and libsnnet was not
doing a lot more than what netlink does. Moreover, libsnnet was not
relying on Handler netlink type to call the funtions, meaning it was
less efficient than this patch. This handler allows to keep the
netlink socket opened, preventing from a lot of open/close on the
same socket through different calls, which is very useful to save
performances.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>