cnm: Fill endpoints results with routes
When using CNM, we were not filling the results of each endpoint
for the Routes field. Now that we can rely on netlink package to
scan the routes inside the specified network namespace, we can
properly fill in this field as expected.

As a comment, this is not needed in case of CNI because the plugin
takes care of this for us.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>