hyperstart: Fix IP netmask and route gateway
Gateway parameter given to hyperstart was sometimes not filled
properly in case it was nil. After a specific check, we can be
sure the gateway value is proper.

With the same idea, the netMask needed to be processed from the
hexadecimal value because hyperstart expect it as an integer.

At last, we don't need NetMask field in the hyperstart structure
HyperNetIface because it is already contained into each structure
HyperIPAddress.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>