pkg: cni: Fix unit testing after CNI plugins update
In order to follow more closely the CNI specification, some recent changes
have been added to the CNI plugins so that they are not failing anymore in
case we try to tear down a non existing network.

This patch fixes the unit test which was expecting to see a failure. Now,
we are expecting a success in removing a network that we've never created
before.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>