pkg: oci: Fix PodConfig filling
In order to get the runtime 3.0 working, we have to fix some fields
in the PodConfig provided by the OCI package.

Here is the list:
- Add missing container ID in ContainerConfig. Make this the same than
pod ID.
- Use the rootfs provided by the config.json instead of an hardcoded
value.
- Add a console path to the Console field of the PodConfig. Otherwise,
the VM will crash because cc-agent.target used as systemd init will
expect to be able to write to that console.

The change about the console expect the runtime configuration to
provide this path.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>