qemu: Device IDs cannot be more than 31 bytes
This patch modifies qemu.go so that we make sure we don't hit the case
where we provide an ID higher than 31 bytes for any new device we ask
to be created by Qemu. We can still hit this case when using indexes,
but this means that we have already created all the previous devices
and there is no remaining room for other devices. This should not
happen.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>