pkg: hyperstart: Export some functions for unit testing
Because the hyperstart package now includes the hyperstart structures
that we need, this commit removes the dependancy on former package
github.com/hyperhq/runv/hyperstart/api/json for our unit tests.

But this is causing a cycle import with the hyperstart mock which
is also used by our unit tests. That's why this commit also takes
care of moving the unit test to its own namespace hyperstart_test
instead of the former hyperstart one. This implies that our unit
tests now need some exported values and some new getters to access
internal hyperstart fields and variables.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>