virtcontainers: Allow logger to be defined by consumer of this package
Virtcontainers is a library/package and there was no way to specify
a log file or a log level for the logrus package used.

This commit adds a new API function to set a specific logrus logger
from the caller. If this new function is not called, the logger will
print to stdout by default.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>