hypervisor: qemu: Add and implement getPodConsole()
In case we run several pods, there is the risk to rely on the same
console path. Because this path is internal, and to prevent from
that case, we add a new function to the hypervisor interface so that
we can actually retrieve this path from anywhere in the library.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>