virtc: Add cc-shim support
In order to properly test hyperstart agent with ccProxy proxy, we
have to be able to start cc-shim to interact with the process running
inside the container, and to get the outputs too.

That's why this patch now allows virtc to start cc-shim with extra
command line options (-cc-shim and -cc-shim-path).

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>