pod: Store proxy URL in pod State
We need the URL to be stored so as to be able to access it anytime
after we first created the Pod. The State structure fit very well
for this usage and that is the reason why this commit extend this
structure with a new URL field.

That way, we can store the URL retrieved from the proxy when doing
the regitration the first time, and we can check the validity of the
URL returned every time the proxy is called with connect().

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>