shim: Introduce new shim interface
We need virtcontainers to be aware about the shim when a shim is
actually needed. Specifically, we are introducing this new interface
in order to comply with cc-proxy which needs a shim to be connected
at the moment we send a command such as starting a new container or
running a new process on an existing container.

We now have to specify which type and which config we want for the
shim, meaning we have to be able to specify this from the PodConfig.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>