virtc: Don't wait for the shim to return
Because we now have the possibility to provide a console for every
new container, we don't need to wait for the end of the shim from
the terminal where we run virtc commands.

Instead we will have another terminal opened to interact with stdin,
stdout and stderr from the container running.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>