qemu: Don't append consoles for every container
Because the consoles with the output from every process started on
every container are managed through the agent/proxy, there is no point
in creating consoles with the Qemu hypervisor.

Moreover, this is inconsistent with the creation of new containers,
leading to the inability of creating new consoles after the VM has
been started.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>