container: Move Console from ContainerConfig to Cmd structure
There is a limitation regarding the ability to provide a different
console than container's one in case we are trying to execute a new
process on this existing container. Indeed, because the console is
tied to the ContainerConfig structure, we have no way to specify a
different console for another process.

This patch moves Interactive and Console fields from the structure
ContainerConfig to Cmd, solving our current limitation.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>