qemu: Prevent udev from creating predictable network interface names
Because we rely on the MAC address to find the network interface, we
don't need to enable the renaming by udev of the interface created.

More than that, because hyperstart agent waits for the interface to be
ready before it can continue, we don't want the name of the interface
found to change after we received the first uevent.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>