OCI/docker: Add Bundle to ContainerConfig and ContainerStatus.
To fully support OCI (or more strictly Docker) compatability,
it is necessary to store the bundle path as that is required
to implement higher-level constructs (such as "$runtime list").

This changes fixes the behavour of StatusToOCIState() which now
returns the Bundle in the State object where previously it
(erroneously) returned the RootFS associated with the bundle.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>