hyperstart: Copy pause binary instead of creating a symbolic link
Initially we used a hardlink to create the pause binary to the needed
location so that our pause container could use it from inside the VM.
But hardlink had a limitation regarding cross partitions, and that's
the reason we moved to symbolic link.

But symbolic links are not evaluated when the pause container rootfs
was mounted, resulting in having the pause binary missing.

The only way to properly deal with those issues is to copy the binary
to the expected location.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>