qemu: Wait for the end of the VM
After we sent the QUIT command through QMP socket to the VM, we were
just returning, assuming the VM was stopped. But in some cases, it
can take a little bit more time for the VM to stop, causing issues
with our benchmarks. Indeed, running the same VM with the same UUID
was conflicting and the QMP sockets were not created because the
previous VM with the same UUID was not stopped.

This patch waits for a go channel indicating QMP socket has been
disconnected, meaning the "real" end of the VM.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>