mounts: Handle bind mounts
We ignore the system mounts for now, as these need to
be created within the VM. We handle the bind-mounts
by first bind-mounting them in the hyperstart shared directory
which is a 9pfs share. We prefix the file-paths with container id
and randomize them to avoid conflicts for same file names
within the container.
These file-paths are made available to hyperstart through the fsmap
structure, so that they can be mounted at the correct location
within the container.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>