state: Add functionality to store/fetch mounts to/from filesystem.
The bind-mounts that are created in the hyperstart shared directory
need to be unmounted when the container is stopped.
We need to store these bind-mount paths created while starting the
container to filesystem, so that they can be retieved later while
deleting the container for performing the unmounts.

Provide convenience functions for storing and retieving mounts
from filesystem for a container.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>