storage: Add function to get the underlying device and mount point
Add helper function in preparation for using devicemapper block
device. This function is used for getting the device major and
minor number for the underlying device containing the file
and the mountpoint of the mount containing the file.

We will be using this for getting the device major-minor associated
with the container rootfs and the mountPoint of the rootfs.

We will eventually use the device to verify if it is a block device
associated with devicemapper.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>