pod: Make state handling functions set in-memory state.
Previously, setPodState() and setContainerState() only set the on-disk
state of a pod or container respectively. They now also set the
in-memory state. This is cleaner, less surprising, removes the burden
from the caller to manage the in-memory state and also ensures that both
versions of the state are kept in sync (safer).

Fixes #279.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>