state: Fix container state update
We plan to add block index and fstype into the container
state. However, when the "STATE" string  in container state
is updated, it resets the entire state struct.
Update the container setContainerState method to fix this behaviour,
such that only the state string is updated.

Refactor the pod.setContainerState function to call the container
method as well instead of the pod updating the container state.
This fixes the behaviour seen above when the pod STATE string
is updated.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>