cnm: Don't ignore routes with empty/nil destination
Because we were ignoring the routes with an empty/nil destination,
we were missing the "default" rule redirecting all the traffic to
the docker0 gateway. This patch takes care of saving routes with
a nil destination now.

Notice that we have to assign a 0.0.0.0 address so that it can be
properly marshalled/unmarshalled. Otherwise, we end up with the
network.json storing "dst":"\u003cnil\u003e", and generating an
error when we try to unmarshal it later.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>