create: Split hyperstart init function to perform setup only at create
The hyperstart agent implementation for init() interface was
performing setup such as adding devices with the hypervisor.
Since init is called during create, start and all other container
lifecycle stages, this setup was being performed repeatedly.

Introduce a new interface for agent called "createPod" that is
supposed to perform all the initial setup while creating a Pod.
Refactor hyperstart init() to exclude the device creation and move
that to createPod.

Create a new function called doFetchPod() that is responsible
for creating the Pod struct. Split createPod() in pod.go to call
doFetchPod() and rest of the functionality that need to be performed
while creating a pod.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>