stop: Wait for end of container process before to remove
In case we are stopping a pod or a container, for every container,
we send a SIGKILL signal and then we try to remove the container.
Unfortunately, if the process inside the container takes some time
to be killed, the removal of the container can sometimes fail.

This patch ensures that we are waiting for the end of the shim
process after we sent the SIGKILL signal. In case we hit a timeout,
the stop() will fail.

Fixes #318

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>