api: Update the container status when retrieving its status
In order to get containers statuses updated when a consumer of this
library asks for those statuses, we have to check the shim process
according to the PID that we have. In case it is not running, we
have to update the status to "stopped" by calling into container
stop() function.

Fixes #327

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>