container: Signal a ready container is valid but not stopping it
This commit removes the ability to stop a container in 'ready' state
but allows a ready container to be killed/signaled. There is no good
reason to allow a ready container to be stopped, but we can be
permissive about "kill". Indeed, if the container is not running and
we try to issue a non-killing signal (i.e different from SIGTERM and
SIGKILL), this will be considered as a no-op. In case of killing
signals, the shim will be killed and the container state will be
set to STOPPED directly.

Fixes #328

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>