pod: Ignore container ready or stopped when stopping the pod
Some containers may have terminated and their status may have been
updated when statusContainer() is called. This should not prevent
the pod to be properly stopped because some containers are already
in "stopped" state or in "ready" state. Instead, the call to pod
stop() will ignore the containers already stopped or not running.

Fixes #331

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>