container: Don't issue an error if trying to stop a stopped container
According to the recent changes allowing to update the container
status if the process has been terminated in the meantime, we could
end up in a situation where the caller wants to stop the container
but the process has terminated before and in that case, we don't
want to issue an error because this means the container has been
stopped earlier. Eventually the result is the same, we have a
stopped container, that's the reason why it should be considered
as a valid case.

Fixes #329

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>